[WIP] add: try-catch

+++ src/main/kotlin/com/example/myHomeAppliance2/AddAppliance.kt
@@ -1,5 +1,5 @@
 package com.example.myHomeAppliance2
 
-data class AddAppliance(val familyId: Long, val applianceName:String ,
-                        val modelNumber: String, val makerId: Long,val categoryName: String,
-                        val usePlaceId: Long, val buyDate: Long ,val buyAt: String)
+data class AddAppliance(val familyId: Int, val applianceName:String ,
+                        val modelNumber: String, val makerId: Int,val categoryName: String,
+                        val usePlaceId: Int, val buyDate: Long ,val buyAt: String)
